requirements.txt
aiogram==3.15.0
python-dotenv==1.0.1
aiosqlite==0.20.0
matplotlib==3.9.0
Pillow==10.4.0
